addToLayaway/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==1.4.2
pymongo==4.2.0
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
